Changed package structure in order to be consistent with "com.kirekov"
