--- src/main/java/com/github/simonharmonicminor/beholder/core/Main.java
@@ -1,7 +1,0 @@
-package com.github.simonharmonicminor.beholder.core;
-
-class Main {
-    public static void main(String[] args) {
-        System.out.println("Hello, world!");
-    }
-}
