--- src/main/java/com/github/simonharmonicminor/beholder/core/ViolationLevel.java
@@ -1,9 +1,0 @@
-package com.github.simonharmonicminor.beholder.core;
-
-/**
- * Represents the violation level of one's verifying
- */
-public enum ViolationLevel {
-    WARN,
-    ERROR,
-}
