--- src/main/java/com/github/simonharmonicminor/beholder/core/configuration/rule/Rule.java
@@ -1,14 +1,0 @@
-package com.github.simonharmonicminor.beholder.core.configuration.rule;
-
-import com.github.simonharmonicminor.beholder.core.ViolationLevel;
-
-/**
- * The base interface for all rules.
- * A rule defines a set of restrictions that should be applied to one code element.
- */
-public interface Rule {
-    /**
-     * @return the level that defines rule importance
-     */
-    ViolationLevel violationLevel();
-}
