--- src/main/java/com/github/simonharmonicminor/beholder/core/configuration/rule/method/ProcedureRule.java
@@ -1,7 +1,0 @@
-package com.github.simonharmonicminor.beholder.core.configuration.rule.method;
-
-/**
- * Represents a rule that defines restrictions to procedures
- */
-public interface ProcedureRule extends MethodRule {
-}
