--- src/main/java/com/github/simonharmonicminor/beholder/core/element/ElementPosition.java
@@ -1,17 +1,0 @@
-package com.github.simonharmonicminor.beholder.core.element;
-
-/**
- * Represents coordinates of one's code element
- */
-public interface ElementPosition {
-
-    /**
-     * @return the line in the source file (1 indexed)
-     */
-    int lineNumber();
-
-    /**
-     * @return the column in the source file (1 indexed)
-     */
-    int columnNumber();
-}
