--- src/main/java/com/github/simonharmonicminor/beholder/core/element/method/FunctionElement.java
@@ -1,11 +1,0 @@
-package com.github.simonharmonicminor.beholder.core.element.method;
-
-/**
- * Represents function declared in one type
- */
-public interface FunctionElement extends MethodElement {
-    /**
-     * @return the type returned by the function
-     */
-    Class<?> returningType();
-}
