--- src/main/java/com/github/simonharmonicminor/beholder/core/element/method/ProcedureElement.java
@@ -1,9 +1,0 @@
-package com.github.simonharmonicminor.beholder.core.element.method;
-
-/**
- * Represents procedure declared in one type.
- * The term "procedure" means that method has been declared
- * with <code>void</code> keyword.
- */
-public interface ProcedureElement extends MethodElement {
-}
