--- src/main/java/com/github/simonharmonicminor/beholder/core/element/type/ClassDeclarationElement.java
@@ -1,8 +1,0 @@
-package com.github.simonharmonicminor.beholder.core.element.type;
-
-
-/**
- * Represents a declared class
- */
-public interface ClassDeclarationElement extends TypeDeclarationElement {
-}
