--- src/main/java/com/github/simonharmonicminor/beholder/core/element/type/EnumDeclarationElement.java
@@ -1,8 +1,0 @@
-package com.github.simonharmonicminor.beholder.core.element.type;
-
-/**
- * Represents enum declaration.
- */
-public interface EnumDeclarationElement extends TypeDeclarationElement {
-
-}
