--- src/main/java/com/github/simonharmonicminor/beholder/core/process/BeholderProcess.java
@@ -1,19 +1,0 @@
-package com.github.simonharmonicminor.beholder.core.process;
-
-import com.github.simonharmonicminor.beholder.core.verifier.VerifyingError;
-import com.github.simonharmonicminor.juu.collection.immutable.ImmutableList;
-
-import java.io.File;
-
-/**
- * Defines main process that scans the project and returns found errors
- */
-public interface BeholderProcess {
-    /**
-     * Verifies given package recursively.
-     *
-     * @param pathToMainPackage path to package that should be verified
-     * @return found errors
-     */
-    ImmutableList<VerifyingError> verifyProject(File pathToMainPackage);
-}
