+++ src/main/java/com/kirekov/beholder/core/Main.java
@@ -1,0 +1,7 @@
+package com.kirekov.beholder.core;
+
+class Main {
+    public static void main(String[] args) {
+        System.out.println("Hello, world!");
+    }
+}
