+++ src/main/java/com/kirekov/beholder/core/ViolationLevel.java
@@ -1,0 +1,9 @@
+package com.kirekov.beholder.core;
+
+/**
+ * Represents the violation level of one's verifying
+ */
+public enum ViolationLevel {
+    WARN,
+    ERROR,
+}
