+++ src/main/java/com/kirekov/beholder/core/configuration/rule/method/FunctionRule.java
@@ -1,0 +1,7 @@
+package com.kirekov.beholder.core.configuration.rule.method;
+
+/**
+ * Represents a rule that defines restrictions to functions
+ */
+public interface FunctionRule extends MethodRule {
+}
