+++ src/main/java/com/kirekov/beholder/core/element/ElementPosition.java
@@ -1,0 +1,17 @@
+package com.kirekov.beholder.core.element;
+
+/**
+ * Represents coordinates of one's code element
+ */
+public interface ElementPosition {
+
+    /**
+     * @return the line in the source file (1 indexed)
+     */
+    int lineNumber();
+
+    /**
+     * @return the column in the source file (1 indexed)
+     */
+    int columnNumber();
+}
