+++ src/main/java/com/kirekov/beholder/core/element/method/MethodElement.java
@@ -1,0 +1,20 @@
+package com.kirekov.beholder.core.element.method;
+
+import com.kirekov.beholder.core.element.Element;
+import com.kirekov.beholder.core.element.param.ParamElement;
+import com.github.simonharmonicminor.juu.collection.immutable.ImmutableList;
+
+/**
+ * The base interface for all procedures and functions declared in one type
+ */
+public interface MethodElement extends Element {
+    /**
+     * @return true if this method overrides one from extended type, returns false otherwise
+     */
+    boolean overridden();
+
+    /**
+     * @return params passed to the method
+     */
+    ImmutableList<ParamElement> methodParams();
+}
