+++ src/main/java/com/kirekov/beholder/core/element/method/ProcedureElement.java
@@ -1,0 +1,9 @@
+package com.kirekov.beholder.core.element.method;
+
+/**
+ * Represents procedure declared in one type.
+ * The term "procedure" means that method has been declared
+ * with <code>void</code> keyword.
+ */
+public interface ProcedureElement extends MethodElement {
+}
