+++ src/main/java/com/kirekov/beholder/core/element/type/ClassDeclarationElement.java
@@ -1,0 +1,8 @@
+package com.kirekov.beholder.core.element.type;
+
+
+/**
+ * Represents a declared class
+ */
+public interface ClassDeclarationElement extends TypeDeclarationElement {
+}
