+++ src/main/java/com/kirekov/beholder/core/element/type/EnumDeclarationElement.java
@@ -1,0 +1,8 @@
+package com.kirekov.beholder.core.element.type;
+
+/**
+ * Represents enum declaration.
+ */
+public interface EnumDeclarationElement extends TypeDeclarationElement {
+
+}
