--- src/test/java/com/github/simonharmonicminor/beholder/core/MainTest.java
@@ -1,13 +1,0 @@
-package com.github.simonharmonicminor.beholder.core;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class MainTest {
-
-    @Test
-    void main() {
-        assertDoesNotThrow(() -> Main.main(new String[]{}));
-    }
-}
