--- src/test/java/com/github/simonharmonicminor/beholder/core/ViolationLevelTest.java
@@ -1,13 +1,0 @@
-package com.github.simonharmonicminor.beholder.core;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-
-class ViolationLevelTest {
-    @Test
-    void testEnumValues() {
-        assertNotNull(ViolationLevel.ERROR);
-        assertNotNull(ViolationLevel.WARN);
-    }
-}
