+++ src/test/java/com/kirekov/beholder/core/MainTest.java
@@ -1,0 +1,13 @@
+package com.kirekov.beholder.core;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class MainTest {
+
+    @Test
+    void main() {
+        assertDoesNotThrow(() -> Main.main(new String[]{}));
+    }
+}
