+++ src/test/java/com/kirekov/beholder/core/ViolationLevelTest.java
@@ -1,0 +1,13 @@
+package com.kirekov.beholder.core;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+
+class ViolationLevelTest {
+    @Test
+    void testEnumValues() {
+        assertNotNull(ViolationLevel.ERROR);
+        assertNotNull(ViolationLevel.WARN);
+    }
+}
